Extract CLI-specific push logic
Lets us reuse the same push strategies across the client without fear of
dependency cycles.
