+++ cachix/src/Cachix/Client/Commands.hs
@@ -11,17 +11,16 @@ module Cachix.Client.Commands
     use,
     remove,
     pin,
-    withPushParams,
   )
 where
 
 import qualified Cachix.API as API
 import Cachix.API.Error
 import Cachix.Client.CNix (filterInvalidStorePath, followLinksToStorePath)
+import Cachix.Client.Commands.Push
 import qualified Cachix.Client.Config as Config
 import Cachix.Client.Env (Env (..))
 import Cachix.Client.Exception (CachixException (..))
-import Cachix.Client.HumanSize (humanSize)
 import qualified Cachix.Client.InstallationMode as InstallationMode
 import qualified Cachix.Client.NixConf as NixConf
 import Cachix.Client.NixVersion (assertNixVersion)
@@ -38,31 +37,23 @@ import Cachix.Client.Secrets
   )
 import Cachix.Client.Servant
 import qualified Cachix.Client.WatchStore as WatchStore
-import qualified Cachix.Types.BinaryCache as BinaryCache
 import qualified Cachix.Types.PinCreate as PinCreate
 import qualified Cachix.Types.SigningKeyCreate as SigningKeyCreate
 import qualified Control.Concurrent.Async as Async
-import Control.Exception.Safe (throwM)
 import Control.Monad.Trans.Maybe (MaybeT (..), runMaybeT)
-import Control.Retry (RetryStatus (rsIterNumber))
 import Crypto.Sign.Ed25519 (PublicKey (PublicKey), createKeypair)
-import qualified Data.ByteString as BS
 import qualified Data.ByteString.Base64 as B64
-import qualified Data.Conduit as Conduit
 import Data.String.Here
 import qualified Data.Text as T
 import qualified Data.Text.IO as T.IO
 import GHC.IO.Handle (hDuplicate, hDuplicateTo)
-import Hercules.CNix.Store (Store, StorePath, storePathToPath, withStore)
-import Network.HTTP.Types (status401, status404)
+import Hercules.CNix.Store (storePathToPath, withStore)
 import Protolude hiding (toS)
 import Protolude.Conv
 import Servant.API (NoContent)
 import Servant.Auth.Client
 import Servant.Client.Streaming
 import Servant.Conduit ()
-import System.Console.AsciiProgress
-import System.Console.Pretty
 import System.Directory (doesFileExist)
 import System.IO (hIsTerminalDevice)
 import qualified System.Posix.Signals as Signals
@@ -113,18 +104,6 @@ IMPORTANT: Make sure to make a backup for the signing key above, as you have the
         Text
     )
 
-notAuthenticatedBinaryCache :: Text -> CachixException
-notAuthenticatedBinaryCache name =
-  AccessDeniedBinaryCache $
-    "Binary cache " <> name <> " doesn't exist or it's private and you need a token: " <> Config.noAuthTokenError
-
-accessDeniedBinaryCache :: Text -> Maybe ByteString -> CachixException
-accessDeniedBinaryCache name maybeBody =
-  AccessDeniedBinaryCache $ "Binary cache " <> name <> " doesn't exist or you don't have access." <> context maybeBody
-  where
-    context Nothing = ""
-    context (Just body) = " Error: " <> toS body
-
 getNixEnv :: IO InstallationMode.NixEnv
 getNixEnv = do
   user <- InstallationMode.getUser
@@ -157,17 +136,6 @@ remove env name = do
   nixEnv <- getNixEnv
   InstallationMode.removeBinaryCache (Config.hostname $ config env) name $
     InstallationMode.getInstallationMode nixEnv InstallationMode.defaultUseOptions
-
-handleCacheResponse :: Text -> Maybe Token -> ClientError -> IO a
-handleCacheResponse name optionalAuthToken err
-  | isErr err status401 && isJust optionalAuthToken = throwM $ accessDeniedBinaryCache name (failureResponseBody err)
-  | isErr err status401 = throwM $ notAuthenticatedBinaryCache name
-  | isErr err status404 = throwM $ BinaryCacheNotFound $ "Binary cache " <> name <> " does not exist."
-  | otherwise = throwM err
-
-failureResponseBody :: ClientError -> Maybe ByteString
-failureResponseBody (FailureResponse _ response) = Just $ toS $ responseBody response
-failureResponseBody _ = Nothing
 
 push :: Env -> PushArguments -> IO ()
 push env (PushPaths opts name cliPaths) = do
@@ -265,90 +233,3 @@ watchExec env pushOpts name cmd args = withPushParams env pushOpts name $ \pushP
   exitWith exitCode
   where
     getProcessHandle (_, _, _, processHandle) = processHandle
-
-pushStrategy :: Store -> Maybe Token -> PushOptions -> Text -> BinaryCache.CompressionMethod -> StorePath -> PushStrategy IO ()
-pushStrategy store authToken opts name compressionMethod storePath =
-  PushStrategy
-    { onAlreadyPresent = pass,
-      on401 = handleCacheResponse name authToken,
-      onError = throwM,
-      onAttempt = \_ _ -> pass,
-      onUncompressedNARStream = showUploadProgress,
-      onDone = pass,
-      Cachix.Client.Push.compressionMethod = compressionMethod,
-      Cachix.Client.Push.compressionLevel = Cachix.Client.OptionsParser.compressionLevel opts,
-      Cachix.Client.Push.omitDeriver = Cachix.Client.OptionsParser.omitDeriver opts
-    }
-  where
-    retryText :: RetryStatus -> Text
-    retryText retryStatus =
-      if rsIterNumber retryStatus == 0
-        then ""
-        else color Yellow $ "retry #" <> show (rsIterNumber retryStatus) <> " "
-
-    showUploadProgress retryStatus size = do
-      let hSize = toS $ humanSize $ fromIntegral size
-      path <- liftIO $ decodeUtf8With lenientDecode <$> storePathToPath store storePath
-
-      isTerminal <- liftIO $ hIsTerminalDevice stdout
-      onTick <-
-        if isTerminal
-          then do
-            let bar = color Blue "[:bar] " <> toS (retryText retryStatus) <> toS path <> " (:percent of " <> hSize <> ")"
-                barLength = T.length $ T.replace ":percent" "  0%" (T.replace "[:bar]" "" (toS bar))
-
-            progressBar <-
-              liftIO $
-                newProgressBar
-                  def
-                    { pgTotal = fromIntegral size,
-                      -- https://github.com/yamadapc/haskell-ascii-progress/issues/24
-                      pgWidth = 20 + barLength,
-                      pgOnCompletion = Just $ color Green "✓ " <> toS path <> " (" <> hSize <> ")",
-                      pgFormat = bar
-                    }
-
-            return $ liftIO . tickN progressBar . BS.length
-          else do
-            -- we append newline instead of putStrLn due to https://github.com/haskell/text/issues/242
-            appendErrText $ retryText retryStatus <> "Pushing " <> path <> " (" <> toS hSize <> ")\n"
-            return $ const pass
-
-      Conduit.awaitForever $ \chunk -> do
-        Conduit.yield chunk
-        onTick chunk
-
-withPushParams :: Env -> PushOptions -> Text -> (PushParams IO () -> IO ()) -> IO ()
-withPushParams env pushOpts name m = do
-  pushSecret <- findPushSecret (config env) name
-  authToken <- Config.getAuthTokenMaybe (config env)
-  compressionMethodBackend <- case pushSecret of
-    PushSigningKey {} -> pure Nothing
-    PushToken {} -> do
-      let token = fromMaybe (Token "") authToken
-      res <- retryHttp $ (`runClientM` clientenv env) $ API.getCache cachixClient token name
-      case res of
-        Left err -> handleCacheResponse name authToken err
-        Right binaryCache -> pure (Just $ BinaryCache.preferredCompressionMethod binaryCache)
-  let compressionMethod = fromMaybe BinaryCache.ZSTD (head $ catMaybes [Cachix.Client.OptionsParser.compressionMethod pushOpts, compressionMethodBackend])
-  withStore $ \store ->
-    m
-      PushParams
-        { pushParamsName = name,
-          pushParamsSecret = pushSecret,
-          pushParamsClientEnv = clientenv env,
-          pushOnClosureAttempt = \full missing -> do
-            unless (null missing) $ do
-              let numMissing = length missing
-                  numCached = length full - numMissing
-              putErrText $ "Pushing " <> show numMissing <> " paths (" <> show numCached <> " are already present) using " <> T.toLower (show compressionMethod) <> " to cache " <> name <> " ⏳\n"
-            return missing,
-          pushParamsStrategy = pushStrategy store authToken pushOpts name compressionMethod,
-          pushParamsStore = store
-        }
-
--- | Put text to stderr without a new line.
---
--- This is safe to use when printing from multiple threads, unlike hPutStrLn which may fail to insert the newline at the right place.
-appendErrText :: (MonadIO m) => Text -> m ()
-appendErrText = hPutStr stderr
